requirements.txt
tensorflow==2.15.0
numpy==1.24.3
streamlit==1.2.0
keras==2.15.0
pillow==10.2.0